+++ examples/04_data.rs
@@ -69,11 +69,7 @@ fn main() {
     // If you goto http://localhost:8080/form and submit the form you should see "Hello, {NAME}"
     // Also goto http://localhost:8080/greet/John and you should see "Hello, John"
 
-    println!(
-        "[04] Listening on http://{}:{}",
-        server.ip_string(),
-        server.port
-    );
+    println!("[04] Listening on http://{}:{}", server.ip, server.port);
 
     // Start the server
     // This will block the current thread